stocks1/requirements.txt
flask
python-dotenv
datetime
requests
pymongo
